+++ src/quickapp/agent/tool_executor.py
@@ -5,7 +5,7 @@ from typing import Dict, List
 
 from injector import inject
 
-from quickapp.agent.chunk_processor import AccumulatedToolCall
+from quickapp.agent._models import AccumulatedToolCall
 from quickapp.common import CompletionResult, StagedBaseTool
 from quickapp.common.perf_timer.perf_timer import PerformanceTimer
 from quickapp.common.utils import sanitize_toolname
